Moved the initialization of the subjectID from initJsPsych to the on_finish segment of the first, fake trial, as it did not seemt to properly work before. Also added some comments to the code.
